+++ generators/src/org/jetbrains/kotlin/generators/tests/GenerateTests.kt
@@ -998,6 +998,10 @@ fun main(args: Array<String>) {
             model("android/rename", recursive = false, extension = null)
         }
 
+        testClass<AbstractAndroidLayoutRenameTest>() {
+            model("android/renameLayout", recursive = false, extension = null)
+        }
+
         testClass<AbstractAndroidFindUsagesTest>() {
             model("android/findUsages", recursive = false, extension = null)
         }
@@ -1157,7 +1161,7 @@ class TestGroup(val testsRoot: String, val testDataRoot: String) {
                     }
                     else {
                         SimpleTestClassModel(rootFile, recursive, excludeParentDirs,
-                                             compiledPattern, filenameStartsLowerCase, testMethod, className, 
+                                             compiledPattern, filenameStartsLowerCase, testMethod, className,
                                              targetBackend, excludeDirs)
                     }
             )