+++ ChangeLog.md
@@ -223,6 +223,7 @@ Issues fixed:
 - [KT-11264](https://youtrack.jetbrains.com/issue/KT-11264) Action to create new activity in Kotlin
 - [KT-11201](https://youtrack.jetbrains.com/issue/KT-11201) Do not ignore items with similar names in kapt
 - [KT-7729](https://youtrack.jetbrains.com/issue/KT-7729) Add Android Lint checks for Kotlin (Android Studio 1.5)
+- Rename Android Extensions imports when the layout file is renamed
 
 ### Maven